feat(home): permanent local-groups mini-section + add 3 press articles
- Replace the conditional 'next action' banner with a local-groups mini-section
  that always shows a short blurb + link to /groupes-locaux, and appends the
  next upcoming action only when one exists (rename next-action -> local-groups-cta)
- Add the 3 missing press-coverage articles to the static fallback (Le Monde
  06/06, Le Nouvel Obs G7 15/06, Channel News 17/06) now that URLs are known
